Removed renderer crap interface.
Overengineering again. Also, using these "renderers" made no sense. I
couldn't even see the progress of the run monitors! My plan was to
output everything to stdout... So now my plan is to add a flag -outfile
and -outtype, which specify the output file and the output type (e.g.
json, csv) respectively. That way I can still see the progress on
stdout, and output the gathered results in a different format.