Refactored the result thing much + PandoraFMS agent support.
Instead of only writing to CSV and JSON, I made a start to create agent data
XML files which can be fed to Pandora. Experimental for a bit.

Also, instead of gathering single results, return a collection of results
which belong to a configuration. This helps with later grouping of results
possible different files.